Move PasscodeSettingsFragment to passcode package
